P2P dialog is dismissed when touched outside
When P2P connection request dialog is shown and user touches
outside the dialog, it will be dismissed. This is not good UX.
For example, tablet has large screen so user may touch outside
unexpectedly. The dialog should be dismissed only when Back key,
Decline button or Accept button is tapped.

Change-Id: Ica93b06190245bd595df67e87bff88e1b4c91e7e