Update patch set 1
Patch Set 1: Code-Review-1

I think we need to do this in a different way .. because the enable/disable doesn't stick all the time. 

An easier way to do this would be to issue disconnect() and reconnect() on WifiStateMachine when SIM is removed/added. Another change we need to do is to check for SIM's presnce if AutoJoinController chose EAP-SIM/AKA/AKA' network as best candidate.

Patch-set: 1
Reviewer: Vinit Deshpande <1022138@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1