Handle all of CMD_SET_OPERATIONAL_MODE in ScanModeState
SCAN_ONLY_WITH_WIFI_OFF_MODE is not handled in ScanModeState. Due to
this implementation WifiStateMachine can not manage the own status
correctly when the device is in idle mode and set as "Scanning Always
Available" on. This commit adds the implementation to handle all of
CMD_SET_OPERATIONAL_MODE include SCAN_ONLY_WITH_WIFI_OFF_MODE and
WifiStateMachine can manage the own state correctly.

Change-Id: If10fb3a7c0ae0c81d8dade3e52f65a8e6f0b1dc6