Close wifi MonitorThread when no more used
During Wifi ON/OFF MTBF tests and in stress conditions, a lot
of Wifi MonitorThread remain active in parallel. This leads to
resource consuming and concurence when treating events coming
from wifi native socket.

As a consequence, on user side, we can observe wifi turning
on/off repeatedly. Sometimes, a normal situation can be
recovered after a couple of minutes but most often, a platform
reboot is necessary. This makes a very bad user experience.

This patch makes sure to close MonitorThread when no interface
uses it anymore. With this patch we have always one and only
one MonitorThread runnning at a time.

Change-Id: I096baf3b5f9d8181c2f83d22d0ed7dfe9514c9da
Signed-off-by: Matthieu MAUGER <matthieux.mauger@intel.com>
Signed-off-by: Abdelmajid MLAYEH <abdelmajidx.mlayeh@intel.com>