Wifi: Fix scan alarm issue
Device that does not support background scan was never reconnected
during screen off. We need to set the alarm status as disabled once
alarm happens so that a new alarm can be enabled.

setScanAlarm() enable parameter is ignored and forced to true if
background scan is not supported. But this support is already tested
by caller functions. moreover, we want to be able to cancel a scan
alarm even if background scan is not supported.

Change-Id: Iaa1356db374c4127539d86ab74232677ff23e98f
Signed-off-by: Loic Poulain <loic.poulain@intel.com>