Close wifi MonitorThread when no more used
During Wifi ON/OFF MTBF tests, a lot of Wifi MonitorThread remain active in
parallel. This leads to resource consuming and concurence when treating events
coming from wifi native socket.

As a consequence, wifi is unable in user side, some times is necesary to reboot
the platform.

This patch makes sure to close MonitorThread when no interface uses it anymore.
Therefore we have always one and only one MonitorThread runnning at a time.

Change-Id: I02ae88546ef600822c7965f7cc82e5c23c05a874
Signed-off-by: Matthieu MAUGER <matthieux.mauger@intel.com>
Signed-off-by: Angel Perez <angel.r.perez@intel.com>
Signed-off-by: Zhiquan Liu <zhiquan.liu@intel.com>