WiFi: Make WiFi scan as the only one scan interval period
When WiFi is disconnected, the screen is on, P2P is not connected
and there are no remembered networks, it is found that two WiFi
scan periods mSupplicantScanIntervalMs and mDisconnectedScanPeriodMs
exist at the same time. The correct logic is when WiFi is disconnected,
the screen is on and P2P is not connected, if there are no remembered
networks, it scans as period mSupplicantScanIntervalMs 15s; If there
are remembered networks, it scans as period mDisconnectedScanPeriodMs
10s. This patch removes the message CMD_NO_NETWORKS_PERIODIC_SCAN to
make the logic more clearer.

Change-Id: I052face1b01c0cb022c8224ad86f14afde5cd767
Signed-off-by: Alice Liu <alice.liu@intel.com>