WiFi: Make WiFi scan as the only one scan interval period
When WiFi is disconnected and the screen is on, it is found
that WiFi scans as two interval periods. when the message
CMD_NO_NETWORKS_PERIODIC_SCAN is processed in DisconnectedState,
it scans as supplicant scan interval 15s. When the message
CMD_START_SCAN is processed in DisconnectedState, it scans
as disconnected scan interval 10s. These two scan interval
periods exist at the same time. The other issue is that when
CMD_START_SCAN message is obsolete, background scan is disabled.
This will cause PNO does not work at that time. This patch
fixes the above two issues.

Change-Id: I052face1b01c0cb022c8224ad86f14afde5cd767
Signed-off-by: Alice Liu <alice.liu@intel.com>