Fix a bug in WifiMonitor
If the Iface is removed when failed to connect to supplicant,
it won't be added again except re-creating the monitor;
however, the monitor only being created when creating wifi/p2p service,
and these 2 service won't be re-created in normal cases.
So, once the Iface is removed, wifi/p2p couldn't be enabled again.

Change-Id: I77f24240508a362ffe17b92ce120d74678485ce8
Signed-off-by: yuwan <yuwan@marvell.com>