Obtain link layer stats from kernel if driver is not loaded
Wi-Fi framework has a mechanism about obtaining link layer statistics
from Wi-Fi driver if it's supported in the driver. If framework can't
obtain it from the driver several times, framework obtains it from
kernel. And framework doesn't have a mechanism to obtain it from the
driver again after obtaining it from kernel.
Since the driver doesn't return it if it's not loaded, there is the
case that framework obtains it from not the driver but kernel even
through it's loaded.

This patch changes the behavior to obtain it from kernel if it's not
loaded, and keeps framework tried to obtain it from the driver.

Bug:30939263

Change-Id: Idcecfda760e20f887229fb36529a0530c68264f1