Fix NullPointerException in AutoJoinController
When WifiAutoJoinController calls WifiNative.status() function it
doesn't check if response is null. Status is a native method
based on doStringCommand that can return null string when supplicant
is not ready, busy, or when it returns "FAIL" keyword.

This patch adds a check as is already done wherever doStringCommand
based methods are called in the framework.

Change-Id: I534a08a7c7f49e799f7d5e6cf08e78732e83a984
Signed-off-by: Matthieu MAUGER <matthieux.mauger@intel.com>
Signed-off-by: liguan1x <guangliangx.li@intel.com>
Signed-off-by: Aymen Ben Alaya <aymenx.ben.alaya@intel.com>
Signed-off-by: Zhiquan Liu <zhiquan.liu@intel.com>