Apply correct default scan interval when P2P is connected
Google recently modified the behavior of WifiStateMachine for the
command "CMD_START_SCAN":
    Author: vandwalle <vandwalle@google.com> 2014-12-10 02:52:35
    Subject: reduce autojoin scan period when P2P is connected
             and infra is disconnected
    Chge-Id: Id1891e6a9f2863d8876915cfefd158fb1712b8fe

However, in case of P2P_connected state they forgot to adjust the
scan interval default value according to this state.

This leads to a lot of undesired scans every 10 seconds while P2P
is connected, reducing dramatically the performances.

This patch fixes that.

Change-Id: Iff504966bc6a25a579e8e61a17be2ed392ddf3c0
Signed-off-by: Matthieu MAUGER <matthieux.mauger@intel.com>
Signed-off-by: Aymen Ben Alaya <aymenx.ben.alaya@intel.com>
Signed-off-by: Zhiquan Liu <zhiquan.liu@intel.com>