Update patch set 1
Patch Set 1: Code-Review-2

This is the purpose of the mNotificationRepeatTime - to prevent us from showing this notification too frequently.  I think the behavior you describe is working as intended.

If you think we shouldn't do that, then you should discuss this with Vinitd@ and the removal would be a larger change.

Patch-set: 1
Reviewer: Robert Greenwalt <1002609@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2