Notify battery stats service when supplicant is killed
Battery History details indicate that Wi-Fi is on when
both "Wi-Fi scanning" and Wi-Fi is disabled. This is due to
battery stats service is not informed that the supplicant has
died.

When supplicant state is reset inform battery stats service
that state is no longer valid. Though to avoid changes in
WifiManager API behavior don't broadcast
SUPPLICANT_STATE_CHANGED_ACTION.

Change-Id: Iab3078b6367e0af5e539f1087722cea94340d62d