Make sure factory reset runs during Wi-Fi on
The function of "Network settings reset(factory reset)" will enable
Wi-Fi and delete all saved networks but it's done at the same time.
This causes all saved networks won't be deleted if Wi-Fi is turned
off since it takes a time to turn on Wi-Fi and all saved network
can't be acquired while Wi-Fi is turned off.
This patch handles factory reset in the state machine to make sure
it's executed while Wi-Fi is turned on.

Bug:31510710

Change-Id: Icb59a983a86d8438c384942bebd15d481692977c