Include WPA_EAP network into the building identity logic
The current code of building identity only handles the networks
that specify IEEE8021X as key management. This doesn't make sense
since not only IEEE8021X but also WPA_EAP should be handled in this
use case. This results in connection failure if the networks have
WPA_EAP only because identity is not built.

Change-Id: I00415552f0cf0f913cded18dc802cb8370b72f27