Update patch set 1
Patch Set 1: Code-Review-2

So I think this has the potential to break some things. In addition to what Vinit said, this will stop the thread on the first event for an interface that is not monitoring. For example if the p2p0 interface is being monitored and wlan0 isn't then if an event comes from wlan0 the monitoring thread will stop even though p2p0 is still listening.

Patch-set: 1
Reviewer: Mitchell Wills <1079753@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2