SoftApManager: move config argument to constructor
This CL moves the softap configuration argument from the start command
to the constructor.  This allows a generic start() method to be used
instead of sending the config with the CMD_START message.  Additionally,
the default config will be used if a config is not provided.  An error
will be triggered if the default config is not loaded properly.

Existing unit tests were updated to reflect the changes above and
additional tests were added to cover a failure loading the default
config.  The tests were also modified to check the SSID name from the
config when start is called.  This allows the test to distinguish the
default from provided configs in the constructor.

Bug: 31631674
Test: manual testing and runtests.sh
Change-Id: I3be2ee36eb8119647fc42623785e1bbd4c7ad9b1