Fix add the duplicate network to wpa_supplicant.conf.
When add a network, if the config allowedKeyManagement
is empty, the key_mgmt in the wpa_supplicant.conf will
not be set, but the result will include NONE,WPA_PSK,
WPA_EAP,IEEE8021X from the supplicant, so the configKey
will be "[ssid]-WPA_PSK", but it is wrong. When add this
network again, it will be duppliate to add because the
"[ssid]-WPA_PSK" is not equal "[ssid]-NONE".

Change-Id: I38feeb782bd7854fa517b05a61500353ed2b62c0
Signed-off-by: zhouzhijie <zhouzhijie@xiaomi.com>