WSM: Handle Disconnect in DisconnectedState
With the new config manager changes, we issue disconnect when the current
network that is connected/connecting to is either removed or disabled.
Since we don't have a "ConnectingState", we need to handle the
disconnect command in the DisconnectedState to ensure that we can cancel
any ongoing wpa_supplicant connection requests.

BUG: 31340625
TEST: Autojoin ACTS test passes now.
Change-Id: Ie4ac75c65ef26f062075404a1916901c4945d243