Update patch set 2
Patch Set 2:

The connecting mode will be added as a part of WifiStateMachine refactor. This was partially reverting one of the fixes that we had made last month for a PNO failure: ag/1253713.

But, it looks like we already fixed this same issue in MR1 branch after that here: ag/1273687. Was this not cherry-picked to aosp during the move?

Patch-set: 2
Reviewer: Roshan Pius <1074459@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0