Update patch set 1
Patch Set 1:

(4 comments)

PTAL - please let me know if I missed any comments. The main update is to clear the cached scan results when we exit the driver started state in the single scan state machine (the preserves the existing behavior in WifiStateMachine).  Tests will be added in the next CL where there will be hooks to get the scan results.  Thanks!

I am also double checking a cherry-pick to master.  Will update when that is done.

Patch-set: 1