WifiService: get scan results from WifiScanner
Retrieve scan results in WifiServiceImpl from WifiScanner instead of
WifiStateMachine.  This utilizes a new method and message type in
WifiScanner.  Tests are added to verify the scan results are returned,
updated and cleared properly.  This CL additionally fixes a few typos
in tests.

Removal of cached scan results from WifiStateMachine will be included in
a separate CL.

BUG: 31444878
TEST: manually tested on angler
TEST: frameworks/opt/net/wifi/tests/wifitests/runtests.sh on angler
TEST: ran GTS CtsNetTestCases on angler
Change-Id: I02c4e24030d2801eb195a84177d908574de0596e