WifiServiceImpl: move scan results dump to scanner
The dump command includes the most recent scan results.  This
information should be dumped directly from WifiScanningServiceImpl
instead of WifiServiceImpl.

The output was also updated to better reflect the age of the entries.
Invalid scan result timestamps are output with "___?___".  Values that
are older than the allotted characters are displayed as ">1000.0".
Valid ages were not changed.  The character representing
isAutoJoinCandidate was removed from the output.  SSID names that are
longer than 32 characters are truncated.

BUG: 31523862
TEST: manual check of bugreport output
TEST: manual check of adb shell dumpsys wifi output
TEST: manual check of adb shell dumpsys wifiscanner output
TEST: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Change-Id: I6529fa7b125a6d7d283fd7dc5f8938632ae4e93f