WiFi Metrics: Add RSSI delta metric (poll v scan)
Created a new metric that logs the difference in RSSI values between a
ScanResult and the first RSSI poll received after attempting to connect
to the AP in that ScanResult.
RSSI poll must come within 3 seconds of frameworks connections attempt,
and is not counted if the connection attempt fails.
The value logged is:
  delta = poll.rssi - scanResult.rssi

Test: Added a unit test
Bug: 31239324
Change-Id: Ice3768df3e8b5c05ff74e7d75e205748403430dc