Update WifiInjector to return a singleton instance
WifiInjector currently has a constructor that creates an instance
of the class, however, the caller needs to know if it has been
created before calling getInstance() method. To address this, this
CL makes the constructor private and passes a context into the method
getInstance(). If the instance doesn't exist, the context argument
will be used to create one, otherwise, it will return the existing
instance. The argument cannot be null, else it will generate a null
pointer exception. Test cases are updated to validate this behavior.

BUG: 31651755
Test: Unit tests suite and power up phone/bring up system service
Change-Id: I16204605e68011177cb1414d153cbfa58c4f0bc3