WifiScoreReportTest: Add unit tests
Tests for WifiScoreReport

Just some very basic tests to start with.
Test that the score is passed to the networkAgent and to the
wifiMetrics.
Make sure the bad linkspeed counter is updated.
Check for saved last report.
Check that the score in the report string matches what was sent out.
Check that the tx and rx rates have influence.

Test: runtests, checkstyle
Bug: 27324715
Change-Id: I9e40f75418af6930260b978c9b6568793e73760b