WifiPermissionsUtil: Resolve refactoring bug
The function canAccessScanResults() changes the semantics of the
permissions check. Specifically, it enforces INTERACT_ACROSS_USERS_FULL
permission on all callers. The original code only enforces this when
the user or profile is not current. Update WifiPermissionsUtil and
WifiPermissionsUtilTest to ensure this functionality. Also, ensure
there's no logical change in the permissions check needed.

Bug: 32400436
Test: Unit test, cherry-pick to master and verify SetupWizard
      can access scan results even though it may not have the
      INTERACT_ACROSS_USERS_FULL permission

Change-Id: I84565946f1a62258b84af0a7b817bcff248d2eda