Add group and pairwise cipher suite info to Capabilities object
This parses the group and pairwise cipher suite from IEs, and stores
them in InformationElementUtil.Capabilities object.

This repaces buildCapabilities() with from() and generateCpabilities
String(). This decouples security capabilities information from the
capabilities string, which is exposed to API and need to match
supplicant format.

This add the pairwise cipher to capabilities string to match the
supplicant format.

This also fixes the WPA/RSN IE structure in comments.

Bug: 32481849
Change-Id: Iaf9ec222886375ed4e5ba66b1d8e0e408acd0a20
Test: compile, manual test, unit tests