Upgrade WPA/EAP connections to their fast-transition equivalent
The developer-visible API and settings should not have to know
whether an AP supports BSS fast transitions, but from a system
point of view it is beneficial (and in some cases necessary)
to take advantage of it where it is possible.  From a config
point of view there is no additional information required
except to enable the authentication type.  This allows us to
handle the feature addition at the lowest layer possible,
assuming that the underlying hardware supports (and is
validated) the fast transition feature.

This implementation will be replaced with one in wificond
once pushing configuration to wpa_supplicant is moved there.

Bug: 32607043
Change-Id: I89a1e0bea3bba3ec8506fc9b239bea8b51c3d45c
Test: Manual: Create an FT-only AP and attempt connections