Update patch set 3
Patch Set 3:

I talked with quiche@ (also added to the review) about the best way to log this.  We should probably be using trace with the new logging.  In parallel, investigate a way to bubble this failure up to the caller.  As of now, the caller cannot distinguish between empty scan results and permissions failures.

Patch-set: 3
Label: Verified=0