WifiStateMachinePrime: activate ScanOnlyMode
Move control for ScanOnlyMode from WifiStateMachine to
WifiStateMachinePrime.  This involves changes to WifiController to
trigger WifiStateMachinePrime to start ScanOnlyMode.  Broadcasts
previously doe in WifiStateMachine are moved to ScanOnlyModeManager.
This additionally removes a transient "flip" of Wifi state where we
previously would briefly broadcast that wifi was enabled - and then
quickly disable - when switching over to scan mode.

Bug: 32661064
Test: runtests.sh
Test: manual testing with toggling wifi and location scans
Change-Id: I89c5e23a36bd4712a3112b691b31502608826bc0