Update patch set 5
Patch Set 5:

(3 comments)

I think it would be really cool to get to short log messages that capture the three types of denial - location permissions, app operations and then profile checks.  Then it is not only easy to search for in the code, but also can adapt as the code changes (if we change how we do a specific part of the check).  it also gives us a good path for throwing exceptions with something useful in them if we can get that in there (which I would really like to do).

Patch-set: 5
Label: Verified=0