ClientModeManager: implement client mode
NOTE: Work-in-progress - not ready for review.  Still need to add tests
for WifiStateMachinePrime Client mode transitions.

Implement the current structure of Client mode in WifiStateMachine in
ClientModeManager.  This will separate the setup and states into
separate phases instead of a long heirarchy of states in
WifiStateMachine.  This CL will not activate the mode, but it will
include tests.

Bug: 32773685
Test: runtests.sh
Change-Id: I8251d261c34cbae789935efe9e8a6fde7f9b93c5