WifiConfigManager: Remove unnecessary loads from store
We are currently loading from store for every user unlock. This is
unnecessary. We want to load from store only in the following scenario:
1. First unlock after device bootup.
2. First unlock after user switch if user was locked during switch.

This should fix the crash in the associated bug where the networkID
changed because of a user unlock (and hence caused an unnecessary store
load previously).

Bug: 32777050
Test: Unit tests
Change-Id: Ie7731900a790e1ecfaf0ad6d7b42addb461b6d40