WifiConfigStore: Move log for store read/write times
The write API in WifiConfigManager is used for both buffered & immediate
writes. So, the log in WifiConfigManager which indicates the time taken
for write for buffered writes is misleading (it's only the time taken
for scheduling the alarm in that case).
So, move it to WifiConfigStore where the eventual write takes place.

Bug: 32778120
Test: Manual tests
Change-Id: I8af5364f735117a1729f737cb476c0f83d447f27