WifiStateMachine: Ignore connection when the network Id is unknown
This is a temporary fix to prevent a NPE crash when the wpa_supplicant
decides to connect to a network on it's own and we've reloaded from
config store and hence changed network Id's.
This will be removed once we migrate to WifiStateMachinePrime.

While there,
Don't try to lookup network configuration in |getTargetSssid| if
|mTargetNetworkId| is -1. This was primarily used for logging in a bunch
of places and it unnecessarily spams the log with
"WifiConfigManager: Looking up network with invalid networkId -1"

Bug: 32834043
Test: Will send for regression testing.
Change-Id: I1f5c6e47f954fc609ca4836c24b4ed241b8aef36