Log API surfaces of WifiScanningService
The following APIs may be used to reach the WifiScanningService
- AIDL methods
- Async Channel messages

Use WifiAsyncChannel instead of AsyncChannel and WifiHandler instead of
Handler to log incoming messages to and outgoing messages from the scanning
service. Since the client establishes a full connection over async channel,
log the messages sent to the client directly by the scanning service.

Bug: 33085782
Test: Unit test suite, verify logging, sanity test (power on and
    connect to Wifi Access Point)

Change-Id: I8df4f4077801c359e11a93fde2c1ce5ea811a0e4