Log API surfaces of WifiScanningService
The following APIs may be used to reach the WifiScanningService
- AIDL methods
- Async Channel messages

Use WifiAsyncChannel instead of AsyncChannel and WifiHandler instead
of Handler in Scanning service to trace messages and add logging
to AIDL methods. WifiAsyncChannel overrides the methods sendMessage()
and replyToMessage() to trace messages seen on the Async Channel.
For fully connected Async Channels, the service doesn't necessary
use the replyToMessage() in AsyncChannel to send messages to the
client. Instead, replies are sent by calling the send() method on
the client's messenger. We add logging to replySucceeded() and to
replyFailed() methods to trace messages sent by the service. Also,
when replyToMessage() method is not used, WifiAsyncChannel is not
enough to log incoming messages. The handler is extended from the
WifiHandler class instead which adds logging of incoming messages to
the service by overriding handleMessage() method.
WifiScanningServiceTest is also updated to mock WifiLog.

Bug: 33085782
Test: Unit test suite, verify logging, sanity test (power on and
    connect to Wifi Access Point)

Change-Id: I8df4f4077801c359e11a93fde2c1ce5ea811a0e4