hotspot2: ANQP elements cleanup Part 4
Cleaned up the implementation and added unit tests for NAIRealmElement
and the classes used by it.

While there, remove the match() function from NAIRealmElement for now.
So that all ANQP elements class are purely data storages for the element.
The match() function will be added to PasspointProvider for matching
a NAI Realm Element against the provider's configuration.

Other changes included:
- removed the now unused EAP.java
- removed the uses of enum for EAP Method ID and Authentication
  Parameter type constants.
- renamed Credential to CredentialType

Bug: 33000864
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Change-Id: I940f6d7cecb9bec405f0786845af768caf527990