Update patch set 2
Patch Set 2:

(5 comments)

You didn't like our messy junk drawer?  ;)

fyi - WifiInjector is having some serious growing pains and we are always debating what we really want it to look like.  One of the issues is that we are dealing with dependencies on other services (ie, p2p) at bad times.  It is a high priority for us to correct these dependencies so we can start WifiService before the other wifi-related services (right now some are before and some after).  Combined with that, not all of our code is refactored to be in a testable state, so this is certainly a work in progress.

Patch-set: 2
Reviewer: Rebecca Silberstein <1081395@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0