hotspot2: add ANQPRequestManager for managing ANQP requests
The main objective is to hold off ANQP requests to an AP
after the previous request is unanwered or failed.

This can eaisly be expanded in the future to throttle ANQP requests
further by taking into consideration of the current connection
status and other criteria, to reduce the impact of ANQP
requests on the network performance and the power consumption.

Bug: 33746564
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Change-Id: I906508f72832a9e461be6b2db2ceacb997765bb8