LogcatLog should use Verbose logging for trace API
Update trace() API in WifiLog.java to use VERBOSE log level instead.
Also, Wifi Verbose Logging setting from Developer settings will be
used along with the log level for the message to determine if it will
be flushed to the log buffer. Expose a new method to set the verbose
logging mode in LogcatLog and synchronize access to the variable that
saves this setting.

Bug: 33783647
Test: Unit test and verify logging
Change-Id: I183f6c9fa148f241b2e9eba9846ccaf34f79ae70