Update patch set 4
Patch Set 4:

(13 comments)

I mostly have requests for an additional item to be logged.  since we want to know who is doing what, can we also have the params traced for a few of these?  (I am thinking of things like setWifiEnabled(boolean).  It would really help to know if an app requested wifi be on or off...  it would make it very clear who is doing what if there is something like an app toggling wifi on while hotspot is active - that shuts off softap :(

Patch-set: 4
Reviewer: Rebecca Silberstein <1081395@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0